requirements.txt
django==1.11.28
certifi==2020.4.5.1
djangorestframework==3.11.0
PyJWT==1.7.1